test(mfsdp): separate overlap test
